Refactor response handling to data layer (repository)
2 little corrections from my side:

1. In this video I mapped the DTO to a model in the domain layer. You should rather do this in the data layer (in the repository)
2. The response handling should also rather be done in the repository (catching HttpException & IOException) as this is specific to Retrofit and the domain layer shouldn't know about that.

Sorry that I didn't show that in the video, but I also never stop learning :)

Q&A
if we do response handling in repository then what the hell use-case do ?
Just everything that's in the use case can be put in the repo and the use case calls the repo function and does nothing else. The use case seems unnecessary then, but it often contains additional business logic like validation or so.